Extract the Claude and Codex hook codecs into cli/integrations/
Why: setup.py carried the full hook-materialization surface alongside the
setup command wiring. Moving the two hook codecs into their own modules
under cli/integrations/ shrinks setup.py and puts the Claude
UserPromptSubmit hook and the Codex lifecycle hooks next to the other
per-surface codecs (json_mcp, codex_config, skills, agents).

This is a pure verbatim relocation with no behavior change:
- cli/integrations/claude_hooks.py: the .claude/settings.json codec
  (HOOK_* constants, materialize_hooks_claude_code, and the add/remove
  helpers), moved character-for-character.
- cli/integrations/codex_hooks.py: the .codex/hooks.json codec
  (_nearest_codex_hooks_repo, _codex_hooks_path, materialize_hooks_codex),
  rendering through the unchanged pure cli/_codex_hooks module, which
  stays in place.

setup.py now imports the three symbols it still calls directly
(materialize_hooks_claude_code, materialize_hooks_codex,
_nearest_codex_hooks_repo) and drops the imports the moved code took with
it (json, find_symlink, atomic_write_text, the nauro_command finders, and
the _codex_hooks render helpers). The two test modules that exercise these
symbols import them from their new locations; assertions are unchanged.
